+++ app/src/main/java/com/fitness/elev8fit/domain/interfaces/authoperations.kt
@@ -2,8 +2,6 @@ package com.fitness.elev8fit.domain.interfaces
 
 import com.fitness.elev8fit.domain.model.Recipe
 import com.fitness.elev8fit.domain.model.User
-import com.fitness.elev8fit.domain.model.chat.ChatRoom
-import com.fitness.elev8fit.domain.model.chat.Message
 import com.fitness.elev8fit.presentation.activity.Recipe.RecipeViewModel
 import com.fitness.elev8fit.presentation.activity.SignUp.SignUpViewModel
 import com.fitness.elev8fit.presentation.activity.socialLoginSignIn.GoogleSignInViewModel
@@ -25,9 +23,5 @@ interface authoperations {
     suspend fun fetchExercises(offset: Int, limit: Int)
 
   suspend fun onNewToken(token: String)
-    suspend fun sendMessageToChatRoom(chatRoomId: String, message: Message)
-    suspend fun fetchChatRoomsForCoach(onChatRoomsFetched: (List<ChatRoom>) -> Unit)
-    suspend fun initializeChatRoom(chatRoomId: String, currentUserId: String)
-    suspend fun observeChatMessages(chatRoomId: String, onMessagesUpdated: (List<Message>) -> Unit)
 
 }
