+++ app/src/main/java/com/fitness/elev8fit/presentation/activity/chat/presentation/ui/ChatRoomListScreen.kt
@@ -1,4 +1,4 @@
-package com.fitness.elev8fit.presentation.activity.chat.ui
+package com.fitness.elev8fit.presentation.activity.chat.presentation.ui
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
@@ -26,7 +26,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.unit.dp
-import com.fitness.elev8fit.presentation.activity.chat.ChatRoomListViewModel
+import com.fitness.elev8fit.presentation.activity.chat.presentation.viewmodel.ChatRoomListViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
