--- app/src/main/java/com/fitness/elev8fit/presentation/intent/ChatroomIntent.kt
@@ -1,6 +1,0 @@
-package com.fitness.elev8fit.presentation.intent
-
-
-sealed class ChatRoomListIntent {
-    object FetchChatRooms : ChatRoomListIntent()
-}
